Set wof:belongsto deterministically
Because we were setting the belongsto properties by iterating over a hash
(which happens in a random order), it was producing a different order of
elements each time, resulting in lots of unneccessary file changes.

This change applies a diff by adding new elements to the end, and dropping any
unneeded elements in place. In addition, it should use the same order for new
elements as the key/value order in wof:hierarchy.