Add ExportChanged function
The current implementation of Export always updates a feature, regardless of
whether anything of meaning has changed, because it increments the lastmodified
timestamp regardless.

However, it's useful sometimes to only perform an export if there are changes
that need writing out, such as when dealing with 2M postcodes, and trying to
minimise the amount of noise git collects.

ExportChanged makes this possible without needing to understand the ins and
outs of preparing and formatting.